requirements.txt
selenium==4.18.1
webdriver-manager==4.0.1
python-dotenv==1.0.1
transformers==4.38.2
torch==2.2.1
pandas==2.2.1
nltk==3.8.1
fake-useragent==1.4.0
requests==2.31.0
aiohttp==3.9.3
python-dateutil==2.8.2
feedparser==6.0.11
pytrends==4.9.2
beautifulsoup4==4.12.3
lxml==5.1.0
html5lib==1.1 